Use Python 3.9 for all basic CI tests
The basic tests do not test the functionality of the code, but rather
runs the developer tools, builds the package distribution and the
documentation.
Since the builds will be done using Python 3.9, this is in alignment.
Note, this does not affect the ability to use Python 3.7 or 3.8 when
installing and using the built package.