Added an additional test for convert (#223)
# Description
Added a test to try to catch the issue seen in SS3_wrappers

## Update
This PR now also fixes an issue in convert. Before
`convert.load_container()` returned a ref. to a blank node if the value
of a recognised key was a dict or list. Now it correctly returns the
list or dict.

A test was also added for this feature.

This PR only fixes a bug and adds more tests. No new functionality is
added.

## Type of change
- [x] Bug fix and code cleanup
- [ ] New feature
- [ ] Documentation update
- [x] Testing


## Checklist for the reviewer
This checklist should be used as a help for the reviewer.

- [ ] Is the change limited to one issue?
- [ ] Does this PR close the issue?
- [ ] Is the code easy to read and understand?
- [ ] Do all new feature have an accompanying new test?
- [ ] Has the documentation been updated as necessary?
- [ ] Is the code properly tested?

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: francescalb <francesca.l.bleken@sintef.no>
Co-authored-by: Francesca L. Bleken <48128015+francescalb@users.noreply.github.com>