Get restrictions as dicts (#225)
# Description
Added a `return_dicts` keyword argument to the
`Triplestore.retrictions()` method, to allow to get the restrictions as
dicts instead of blank nodes.

Other changes:
* Also renamed the `target` argument of `Triplestore.add_retriction()`
and `Triplestore.retrictions()` to `value` in order to be consistent
with the naming in EMMOntoPy. These two methods are a very new addition
to tripper that haven't been used yet. This change that this change will
break existing code is therefore very small.
* Changed argument name and default value of `Triplestore.restrictions() from 
  `return_dicts=False` to `asdict=True`.
* Added tests.
* Added documentation in the tutorial.

## Type of change
- [x] Bug fix and code cleanup
- [x] New feature
- [x] Documentation update
- [x] Testing

---------

Co-authored-by: Francesca L. Bleken <48128015+francescalb@users.noreply.github.com>