Avoid that tripper.Namespace crashes if the cache directory cannot be accessed (#235)
# Description
Avoid that tripper.Namespace crashes if the cache directory cannot be
accessed.

Unfortunately, if you use tripper with the rdflib backend, you may still
get errors if the cache dir is not accessible. Try e.g.

    XDG_CACHE_HOME=/root/.cache pytest tests/test_namespace.py

as a non-root user. It will result in a `HTTPError`, which is caused by
that rdflib doesn't have permissions to store a downloaded turtle file
in the cache directory.

## Type of change
- [x] Bug fix and code cleanup
- [ ] New feature
- [ ] Documentation update
- [ ] Testing


## Checklist for the reviewer
This checklist should be used as a help for the reviewer.

- [ ] Is the change limited to one issue?
- [ ] Does this PR close the issue?
- [ ] Is the code easy to read and understand?
- [ ] Do all new feature have an accompanying new test?
- [ ] Has the documentation been updated as necessary?
- [ ] Is the code properly tested?