Dataset (#256)
# Description
Added new `tripper.dataset` module for documenting datasets.

All the main features documented in the module docstring have now been
implemented.

For the next PRs, we should:
- [ ] Add some kind of session with a user-specific information that
should not be stored in the triplestore, including:
      - credentials like username and password to external storages
      - local root for relative paths
- default storage (e.g. redis) to work against during execution of the
user workflow
Exactly how this session should work is still unclear. It will probably
be a class with the functions documented in the
top of the dataset module as methods + some additional methods for
dealing with the internal state.
- [ ] Decide whether we want to move the protocol plugins from DLite to
Tripper? It may also be possible to create a new package.
- [ ] Update CONTEXT_URL to point to master branch (or to github pages
if we copy the context there)
- [ ] Create a good example/tutorial
- [ ] Document the `@context`
- [ ] Fix the `__TODO__`'s in the code


## Type of change
- [ ] Bug fix and code cleanup
- [x] New feature
- [ ] Documentation update
- [ ] Testing


## Checklist for the reviewer
This checklist should be used as a help for the reviewer.

- [ ] Is the change limited to one issue?
- [ ] Does this PR close the issue?
- [ ] Is the code easy to read and understand?
- [ ] Do all new feature have an accompanying new test?
- [ ] Has the documentation been updated as necessary?
- [ ] Is the code properly tested?

---------

Co-authored-by: Francesca L. Bleken <48128015+francescalb@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Tor S. Haugland <torshaugland@gmail.com>