Add resources (classes) to the JSON-LD context. (#366)
From the data documentation in COLD we learned that it is useful to also include the classes in the JSON-LD context.

This makes it e.g. possible to short names when referring to classes in JSON-LD files. Ex: `{"@type": "Resource", ...}`

This PR is also used in PR #360 to replace keywords with JSON-LD context in many parts of the dataset API. The benefit with this is that this allows to use existing externally defined JSON-LD contexts.

Also added a `__dir__()` method to Keywords that include the documented attributes. It simplifies interactive exploration of Keyword instances using ipython (used while implementing this PR).