Added units submodule for working with units and quantities defined in ontologies (#324)
# Description
Added a new units submodule for working with units and quantities
defined in ontologies.

The module provides support to populate a Pint units definition file
from EMMO. It also subclasses `pint.UnitRegistry`, `pint.Unit`, and
`pint.Quantity`, enhancing them with knowledge from the ontology.

Currently it only supports EMMO as source. But it utilises the
references to QUDT and OM such that it is possible to work with IRIs to
these ontologies as well.

The PR includes tests and documentation.

## Other fixes:
- Added support for quantities (emmo:SIQuantityDatatype) in
`literals.py`
- Silence warning from doctest in `literal.py`
- More informative error message from `Triplestore.value()`
- Corrected type annotation of `Triplestore.restrictions()`
- Made `AttrDict` pickleable
- Run mkdocs and other tools on GitHub with Python 3.11, since it Python
3.9 doesn't understand TypeAlias
- Added `test_markdown_doctest` to run doctest on markdown documentation
in the `docs/` folder

## Follow-up tasks
- [x] Allow to update the unit registry from domain or application
ontologies.
- [x] Given IRI to a quantity in the KB, represent it as a Quantity
object.
- [x] Store a Quantity object to KB.
- [x] Load a Quantity object to KB.
- [ ] Add physical constants from the ontology. 
> Make no sense before physical constants are defined with a value in
EMMO (requires that we deal with authority (and preferable also with
uncertainties))
- [x] Add documentation