Development branch that merges loose ends together (#380)
# Description
We are in a state with many open PRs with mutual conflicts. This branch
merge them together and resolve conflicts.
It merges the following branches:

- context-module, PR #378
- keyword_bugs, PR #375
- master
- matchmaker-demo, PR #377
- fix-readme-link, PR #379
- use_pyld, PR #376

Furthermore it has added support for the graph keyword (in a better
way), replacing PR #360 as well.
See [Tripper network](https://github.com/EMMC-ASBL/tripper/network) for
details.

The speed of tabledoc has been greatly improved. The PINK example with
1021 datasets now runs in a few seconds compared to minutes before...

---------

Co-authored-by: Francesca.L.Bleken@sintef.no <francesca.l.bleken@sintef.no>