Added available() method and check_url to argument to Triplestore (#437)
# Description
Integrated checking if a backend service is available to the standard
Triplestore API, by adding the available() method and check_url to
argument to the Triplestore class.

The `check_url` configuration can now be configured together with other
parameters in the session.

Simplified tests by using sessions and the new `check_url`
configuration.

## Type of change
- [ ] Bug fix and code cleanup
- [x] New feature
- [ ] Documentation update
- [ ] Testing

---------

Co-authored-by: Francesca L. Bleken <48128015+francescalb@users.noreply.github.com>