Fixed some issues with sparql backend (#453)
# Description
Found a few issues in the sparqlwrapper backend when testing SELECT and
UPDATE sparql queries against the MatCHMaker GraphDB instance.

Not sure why these issues haven't been detected before. SELECT and
UPDATE queries are well tested in the
test_sparqlwrapper_graphdb_fusiki.py. Maybe it depends on the version of
the GraphDB instance?