Allow keyword redefinition (#465)
# Description
Desired behaviour:

1. 
It should be allowed to redfine keywords locally. This means that
keywords in the default are overwritten. For instance: vcard:hasName can
be overwritten to fam:hasName.

This behaviour is needed, but should be treated with care. 

We use the argument `redefine` with thre possible values:
'true' (warning issued), 'skip' (warning issued), 'raise'  (default)

Default `raise`: Raises an error to make sure that the choice of
allowing redefinition of or skipping new of keywords is always an active
choice .

In practice this argument will often be set to true or skip, but by
having the default False it should reduce the risk of doing it without
intending to.

'true': existing keywords will be redefined, i.e. overwritten in the
local context.

'skip': if a keyword that already is set is tried to be defined it will
be skipped. This will typically happen if you have a prefLabel in your
ontology that is the same as an existing keyword. With skip, this
keyword will stay as the original. It should issue a warning and say
that you should not use the keyword/preflabel alone for documentating
this property/class but the curie.

---------

Co-authored-by: Jesper Friis <jesper.friis@sintef.no>