Speedup (#474)
# Description
Avoid unnecessary loading har parsing of existing keywords.

This provides a general speed improvement. For Keywords.load_rdf() this
provides a large speedup, since we only load the default keywords once,
instead of for each keyword in the rdf source.

## Type of change
- [x] Bug fix and code cleanup
- [ ] New feature
- [ ] Documentation update
- [ ] Testing


## Checklist for the reviewer
This checklist should be used as a help for the reviewer.

- [ ] Is the change limited to one issue?
- [ ] Does this PR close the issue?
- [ ] Is the code easy to read and understand?
- [ ] Do all new feature have an accompanying new test?
- [ ] Has the documentation been updated as necessary?
- [ ] Is the code properly tested?