Make it possible to duplicate nested table headers (#467)
# Description
Make it possible to duplicate nested table headers.

Currently you can e.g. not represent a dataset with multiple
distributions with TableDoc. This PR allows that by adding brackets as
prefix. You can now have the following header:

| @id | @type | @type | distribution[1].downloadURL |
distribution[1].mediaType | distribution[2].downloadURL |
|---|---|---|---|---|---|

which will create two distributions, the first with both a `downloadURL`
and `mediaType` and the second with only a `downloadURL`.

Brackets can also be used to make the `@type` header fields unique:

| @id | @type[1] | @type[2] | distribution[1].downloadURL |
distribution[1].mediaType | distribution[2].downloadURL |
|---|---|---|---|---|---|

This may be useful if the table e.g. comes from pandas that requires
that all reader fields are unique.


This PR tries to address the needs by datadocweb.

## Type of change
- [ ] Bug fix and code cleanup
- [x] New feature
- [ ] Documentation update
- [ ] Testing


## Checklist for the reviewer
This checklist should be used as a help for the reviewer.

- [ ] Is the change limited to one issue?
- [ ] Does this PR close the issue?
- [ ] Is the code easy to read and understand?
- [ ] Do all new feature have an accompanying new test?
- [ ] Has the documentation been updated as necessary?
- [ ] Is the code properly tested?